requirements.txt
asgiref==3.7.2
Django==5.0
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
